+++ CHANGELOG.md
@@ -5,9 +5,10 @@ MINOR
 ### CryptoOnramp (Alpha)
 * [Added] Added API models for upcoming EU support. 
 * [Added] Added `birthCountry`, `birthCity`, and `nationalities` properties to `KycInfo`, required for EU customers.
+* [Added] Added `CryptoOnrampCoordinator` APIs for EU support: `retrieveMissingIdentifiers()`, `submitIdentifiers(_:)`, and `presentCRSCARFDeclaration(from:)`.
 
 ### PaymentSheet
-* [Added] Added support for new and upcoming payment methods in native Link, such as UPI, Pix, and Crypto
+* [Fixed] Changed `@_spi(STP)` exposure of `LinkAppearance` and `VerifyKYCResult` to `@_spi(CryptoOnrampAlpha)`.
 
 ## 25.13.0 2026-05-04
 ### PaymentSheet